Create Client_hello_v1.1
Choice block needs to be changed and more cipher specs to be added to list.

Missing proper relations. Peach relation does not allow value of an element (number) to be taken as size of another element. What it allows is complete opposite where an elements value can be the size of another element.