Create simplified_client_grammar
This grammar is used to solve an error that occurs with AFLSmart. This grammar crashes with a memory error when its run with AFLSmart.

Removing "<Number name="Sequence_Number" size="48" signed="false" endian="Big"/>" from the first block makes the grammar run fine. Note: even after this change it may crash but runs fine when we run it again.